@@ -2,20 +2,20 @@ package com.auth0.jwt.impl;
 
 import com.auth0.jwt.interfaces.Claim;
 import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.ObjectReader;
 import com.fasterxml.jackson.databind.node.TextNode;
 import org.hamcrest.collection.IsCollectionWithSize;
-import org.hamcrest.core.IsCollectionContaining;
+import org.hamcrest.core.IsIterableContaining;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
-import org.mockito.Mockito;
 
-import java.util.Collections;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
+import java.time.Instant;
+import java.util.*;
 
+import static com.auth0.jwt.impl.JWTParser.getDefaultObjectMapper;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
 
@@ -25,142 +25,156 @@ public class PayloadImplTest {
     public ExpectedException exception = ExpectedException.none();
 
     private PayloadImpl payload;
-    private Date expiresAt;
-    private Date notBefore;
-    private Date issuedAt;
+    private final Instant expiresAt = Instant.now().plusSeconds(10);
+    private final Instant notBefore = Instant.now();
+    private final Instant issuedAt = Instant.now();
+
+    private ObjectMapper objectMapper;
 
     @Before
-    public void setUp() throws Exception {
-        expiresAt = Mockito.mock(Date.class);
-        notBefore = Mockito.mock(Date.class);
-        issuedAt = Mockito.mock(Date.class);
+    public void setUp() {
+        objectMapper = getDefaultObjectMapper();
+
         Map<String, JsonNode> tree = new HashMap<>();
         tree.put("extraClaim", new TextNode("extraValue"));
-        payload = new PayloadImpl("issuer", "subject", Collections.singletonList("audience"), expiresAt, notBefore, issuedAt, "jwtId", tree);
+        payload = new PayloadImpl("issuer", "subject", Collections.singletonList("audience"), expiresAt, notBefore, issuedAt, "jwtId", tree, objectMapper);
     }
 
-    @SuppressWarnings("Convert2Diamond")
     @Test
-    public void shouldHaveUnmodifiableTree() throws Exception {
+    public void shouldHaveUnmodifiableTree() {
         exception.expect(UnsupportedOperationException.class);
-        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, new HashMap<String, JsonNode>());
+        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, new HashMap<>(), objectMapper);
         payload.getTree().put("something", null);
     }
 
     @Test
-    public void shouldGetIssuer() throws Exception {
+    public void shouldHaveUnmodifiableAudience() {
+        exception.expect(UnsupportedOperationException.class);
+        PayloadImpl payload = new PayloadImpl(null, null, new ArrayList<>(), null, null, null, null, null, objectMapper);
+        payload.getAudience().add("something");
+    }
+
+    @Test
+    public void shouldGetIssuer() {
         assertThat(payload, is(notNullValue()));
         assertThat(payload.getIssuer(), is("issuer"));
     }
 
     @Test
-    public void shouldGetNullIssuerIfMissing() throws Exception {
-        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null);
+    public void shouldGetNullIssuerIfMissing() {
+        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null, objectMapper);
         assertThat(payload, is(notNullValue()));
         assertThat(payload.getIssuer(), is(nullValue()));
     }
 
     @Test
-    public void shouldGetSubject() throws Exception {
+    public void shouldGetSubject() {
         assertThat(payload, is(notNullValue()));
         assertThat(payload.getSubject(), is("subject"));
     }
 
     @Test
-    public void shouldGetNullSubjectIfMissing() throws Exception {
-        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null);
+    public void shouldGetNullSubjectIfMissing() {
+        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null, objectMapper);
         assertThat(payload, is(notNullValue()));
         assertThat(payload.getSubject(), is(nullValue()));
     }
 
     @Test
-    public void shouldGetAudience() throws Exception {
+    public void shouldGetAudience() {
         assertThat(payload, is(notNullValue()));
 
         assertThat(payload.getAudience(), is(IsCollectionWithSize.hasSize(1)));
-        assertThat(payload.getAudience(), is(IsCollectionContaining.hasItems("audience")));
+        assertThat(payload.getAudience(), is(IsIterableContaining.hasItems("audience")));
     }
 
     @Test
-    public void shouldGetNullAudienceIfMissing() throws Exception {
-        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null);
+    public void shouldGetNullAudienceIfMissing() {
+        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null, objectMapper);
         assertThat(payload, is(notNullValue()));
         assertThat(payload.getAudience(), is(nullValue()));
     }
 
     @Test
-    public void shouldGetExpiresAt() throws Exception {
+    public void shouldGetExpiresAt() {
         assertThat(payload, is(notNullValue()));
-        assertThat(payload.getExpiresAt(), is(expiresAt));
+        assertThat(payload.getExpiresAt(), is(Date.from(expiresAt)));
+        assertThat(payload.getExpiresAtAsInstant(), is(expiresAt));
     }
 
     @Test
-    public void shouldGetNullExpiresAtIfMissing() throws Exception {
-        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null);
+    public void shouldGetNullExpiresAtIfMissing() {
+        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null, objectMapper);
         assertThat(payload, is(notNullValue()));
         assertThat(payload.getExpiresAt(), is(nullValue()));
+        assertThat(payload.getExpiresAtAsInstant(), is(nullValue()));
     }
 
     @Test
-    public void shouldGetNotBefore() throws Exception {
+    public void shouldGetNotBefore() {
         assertThat(payload, is(notNullValue()));
-        assertThat(payload.getNotBefore(), is(notBefore));
+        assertThat(payload.getNotBefore(), is(Date.from(notBefore)));
+        assertThat(payload.getNotBeforeAsInstant(), is(notBefore));
     }
 
     @Test
-    public void shouldGetNullNotBeforeIfMissing() throws Exception {
-        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null);
+    public void shouldGetNullNotBeforeIfMissing() {
+        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null, objectMapper);
         assertThat(payload, is(notNullValue()));
         assertThat(payload.getNotBefore(), is(nullValue()));
+        assertThat(payload.getNotBeforeAsInstant(), is(nullValue()));
     }
 
     @Test
-    public void shouldGetIssuedAt() throws Exception {
+    public void shouldGetIssuedAt() {
         assertThat(payload, is(notNullValue()));
-        assertThat(payload.getIssuedAt(), is(issuedAt));
+        assertThat(payload.getIssuedAt(), is(Date.from(issuedAt)));
+        assertThat(payload.getIssuedAtAsInstant(), is(issuedAt));
     }
 
     @Test
-    public void shouldGetNullIssuedAtIfMissing() throws Exception {
-        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null);
+    public void shouldGetNullIssuedAtIfMissing() {
+        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null, objectMapper);
         assertThat(payload, is(notNullValue()));
         assertThat(payload.getIssuedAt(), is(nullValue()));
+        assertThat(payload.getIssuedAtAsInstant(), is(nullValue()));
     }
 
     @Test
-    public void shouldGetJWTId() throws Exception {
+    public void shouldGetJWTId() {
         assertThat(payload, is(notNullValue()));
         assertThat(payload.getId(), is("jwtId"));
     }
 
     @Test
-    public void shouldGetNullJWTIdIfMissing() throws Exception {
-        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null);
+    public void shouldGetNullJWTIdIfMissing() {
+        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null, objectMapper);
         assertThat(payload, is(notNullValue()));
         assertThat(payload.getId(), is(nullValue()));
     }
 
     @Test
-    public void shouldGetExtraClaim() throws Exception {
+    public void shouldGetExtraClaim() {
         assertThat(payload, is(notNullValue()));
         assertThat(payload.getClaim("extraClaim"), is(instanceOf(JsonNodeClaim.class)));
         assertThat(payload.getClaim("extraClaim").asString(), is("extraValue"));
     }
 
     @Test
-    public void shouldGetNotNullExtraClaimIfMissing() throws Exception {
-        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null);
+    public void shouldGetNotNullExtraClaimIfMissing() {
+        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, null, objectMapper);
         assertThat(payload, is(notNullValue()));
         assertThat(payload.getClaim("missing"), is(notNullValue()));
-        assertThat(payload.getClaim("missing"), is(instanceOf(NullClaim.class)));
+        assertThat(payload.getClaim("missing").isMissing(), is(true));
+        assertThat(payload.getClaim("missing").isNull(), is(false));
     }
 
     @Test
-    public void shouldGetClaims() throws Exception {
+    public void shouldGetClaims() {
         Map<String, JsonNode> tree = new HashMap<>();
         tree.put("extraClaim", new TextNode("extraValue"));
         tree.put("sub", new TextNode("auth0"));
-        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, tree);
+        PayloadImpl payload = new PayloadImpl(null, null, null, null, null, null, null, tree, objectMapper);
         assertThat(payload, is(notNullValue()));
         Map<String, Claim> claims = payload.getClaims();
         assertThat(claims, is(notNullValue()));
@@ -170,7 +184,7 @@ public class PayloadImplTest {
     }
 
     @Test
-    public void shouldNotAllowToModifyClaimsMap() throws Exception {
+    public void shouldNotAllowToModifyClaimsMap() {
         assertThat(payload, is(notNullValue()));
         Map<String, Claim> claims = payload.getClaims();
         assertThat(claims, is(notNullValue()));
