@@ -2,6 +2,8 @@ package com.auth0.jwt.algorithms;
 
 import com.auth0.jwt.exceptions.SignatureGenerationException;
 import com.auth0.jwt.exceptions.SignatureVerificationException;
+import com.auth0.jwt.interfaces.DecodedJWT;
+import java.util.Base64;
 
 class NoneAlgorithm extends Algorithm {
 
@@ -10,10 +12,21 @@ class NoneAlgorithm extends Algorithm {
     }
 
     @Override
-    public void verify(byte[] contentBytes, byte[] signatureBytes) throws SignatureVerificationException {
-        if (signatureBytes.length > 0) {
-            throw new SignatureVerificationException(this);
+    public void verify(DecodedJWT jwt) throws SignatureVerificationException {
+        try {
+            byte[] signatureBytes = Base64.getUrlDecoder().decode(jwt.getSignature());
+
+            if (signatureBytes.length > 0) {
+                throw new SignatureVerificationException(this);
+            }
+        } catch (IllegalArgumentException e) {
+            throw new SignatureVerificationException(this, e);
         }
+    }
+
+    @Override
+    public byte[] sign(byte[] headerBytes, byte[] payloadBytes) throws SignatureGenerationException {
+        return new byte[0];
     }
 
     @Override
