@@ -2,12 +2,15 @@ package com.auth0.jwt.impl;
 
 import com.auth0.jwt.exceptions.JWTDecodeException;
 import com.auth0.jwt.interfaces.Claim;
+import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.core.ObjectCodec;
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.io.IOException;
 import java.lang.reflect.Array;
+import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -18,35 +21,42 @@ import java.util.Map;
  */
 class JsonNodeClaim implements Claim {
 
+    private final ObjectCodec codec;
     private final JsonNode data;
 
-    private JsonNodeClaim(JsonNode node) {
+    private JsonNodeClaim(JsonNode node, ObjectCodec codec) {
         this.data = node;
+        this.codec = codec;
     }
 
     @Override
     public Boolean asBoolean() {
-        return !data.isBoolean() ? null : data.asBoolean();
+        return isMissing() || isNull() || !data.isBoolean() ? null : data.asBoolean();
     }
 
     @Override
     public Integer asInt() {
-        return !data.isNumber() ? null : data.asInt();
+        return isMissing() || isNull() || !data.isNumber() ? null : data.asInt();
+    }
+
+    @Override
+    public Long asLong() {
+        return isMissing() || isNull() || !data.isNumber() ? null : data.asLong();
     }
 
     @Override
     public Double asDouble() {
-        return !data.isNumber() ? null : data.asDouble();
+        return isMissing() || isNull() || !data.isNumber() ? null : data.asDouble();
     }
 
     @Override
     public String asString() {
-        return !data.isTextual() ? null : data.asText();
+        return isMissing() || isNull() || !data.isTextual() ? null : data.asText();
     }
 
     @Override
     public Date asDate() {
-        if (!data.canConvertToLong()) {
+        if (isMissing() || isNull() || !data.canConvertToLong()) {
             return null;
         }
         long seconds = data.asLong();
@@ -54,55 +64,95 @@ class JsonNodeClaim implements Claim {
     }
 
     @Override
+    public Instant asInstant() {
+        if (isMissing() || isNull() || !data.canConvertToLong()) {
+            return null;
+        }
+        long seconds = data.asLong();
+        return Instant.ofEpochSecond(seconds);
+    }
+
+    @Override
     @SuppressWarnings("unchecked")
-    public <T> T[] asArray(Class<T> tClazz) throws JWTDecodeException {
-        if (!data.isArray()) {
+    public <T> T[] asArray(Class<T> clazz) throws JWTDecodeException {
+        if (isMissing() || isNull() || !data.isArray()) {
             return null;
         }
 
-        ObjectMapper mapper = new ObjectMapper();
-        T[] arr = (T[]) Array.newInstance(tClazz, data.size());
+        T[] arr = (T[]) Array.newInstance(clazz, data.size());
         for (int i = 0; i < data.size(); i++) {
             try {
-                arr[i] = mapper.treeToValue(data.get(i), tClazz);
+                arr[i] = codec.treeToValue(data.get(i), clazz);
             } catch (JsonProcessingException e) {
-                throw new JWTDecodeException("Couldn't map the Claim's array contents to " + tClazz.getSimpleName(), e);
+                throw new JWTDecodeException("Couldn't map the Claim's array contents to " + clazz.getSimpleName(), e);
             }
         }
         return arr;
     }
 
     @Override
-    public <T> List<T> asList(Class<T> tClazz) throws JWTDecodeException {
-        if (!data.isArray()) {
+    public <T> List<T> asList(Class<T> clazz) throws JWTDecodeException {
+        if (isMissing() || isNull() || !data.isArray()) {
             return null;
         }
 
-        ObjectMapper mapper = new ObjectMapper();
         List<T> list = new ArrayList<>();
         for (int i = 0; i < data.size(); i++) {
             try {
-                list.add(mapper.treeToValue(data.get(i), tClazz));
+                list.add(codec.treeToValue(data.get(i), clazz));
             } catch (JsonProcessingException e) {
-                throw new JWTDecodeException("Couldn't map the Claim's array contents to " + tClazz.getSimpleName(), e);
+                throw new JWTDecodeException("Couldn't map the Claim's array contents to " + clazz.getSimpleName(), e);
             }
         }
         return list;
     }
 
     @Override
-    public <T> T as(Class<T> tClazz) throws JWTDecodeException {
-        ObjectMapper mapper = new ObjectMapper();
-        try {
-            return mapper.treeAsTokens(data).readValueAs(tClazz);
+    public Map<String, Object> asMap() throws JWTDecodeException {
+        if (isMissing() || isNull() || !data.isObject()) {
+            return null;
+        }
+
+        TypeReference<Map<String, Object>> mapType = new TypeReference<Map<String, Object>>() {
+        };
+
+        try (JsonParser parser = codec.treeAsTokens(data)) {
+            return parser.readValueAs(mapType);
         } catch (IOException e) {
-            throw new JWTDecodeException("Couldn't map the Claim value to " + tClazz.getSimpleName(), e);
+            throw new JWTDecodeException("Couldn't map the Claim value to Map", e);
+        }
+    }
+
+    @Override
+    public <T> T as(Class<T> clazz) throws JWTDecodeException {
+        try {
+            if (isMissing() || isNull()) {
+                return null;
+            }
+            return codec.treeToValue(data, clazz);
+        } catch (JsonProcessingException e) {
+            throw new JWTDecodeException("Couldn't map the Claim value to " + clazz.getSimpleName(), e);
         }
     }
 
     @Override
     public boolean isNull() {
-        return !(data.isArray() || data.canConvertToLong() || data.isTextual() || data.isNumber() || data.isBoolean());
+        return !isMissing() && data.isNull();
+    }
+
+    @Override
+    public boolean isMissing() {
+        return data == null || data.isMissingNode();
+    }
+
+    @Override
+    public String toString() {
+        if (isMissing()) {
+            return "Missing claim";
+        } else if (isNull()) {
+            return "Null claim";
+        }
+        return data.toString();
     }
 
     /**
@@ -110,23 +160,23 @@ class JsonNodeClaim implements Claim {
      *
      * @param claimName the Claim to search for.
      * @param tree      the JsonNode tree to search the Claim in.
+     * @param objectCodec the object codec in use for deserialization
      * @return a valid non-null Claim.
      */
-    static Claim extractClaim(String claimName, Map<String, JsonNode> tree) {
+    static Claim extractClaim(String claimName, Map<String, JsonNode> tree, ObjectCodec objectCodec) {
         JsonNode node = tree.get(claimName);
-        return claimFromNode(node);
+        return claimFromNode(node, objectCodec);
     }
 
     /**
      * Helper method to create a Claim representation from the given JsonNode.
      *
      * @param node the JsonNode to convert into a Claim.
+     * @param objectCodec the object codec in use for deserialization
      * @return a valid Claim instance. If the node is null or missing, a NullClaim will be returned.
      */
-    static Claim claimFromNode(JsonNode node) {
-        if (node == null || node.isNull() || node.isMissingNode()) {
-            return new NullClaim();
-        }
-        return new JsonNodeClaim(node);
+    static Claim claimFromNode(JsonNode node, ObjectCodec objectCodec) {
+        return new JsonNodeClaim(node, objectCodec);
     }
+
 }
