@@ -1,20 +1,25 @@
 package com.auth0.jwt.algorithms;
 
+import com.auth0.jwt.JWT;
 import com.auth0.jwt.exceptions.SignatureGenerationException;
 import com.auth0.jwt.exceptions.SignatureVerificationException;
-import org.apache.commons.codec.binary.Base64;
+import com.auth0.jwt.interfaces.DecodedJWT;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
+import java.io.ByteArrayOutputStream;
+import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
-import java.util.Arrays;
 
+import static com.auth0.jwt.algorithms.CryptoTestHelper.asJWT;
+import static com.auth0.jwt.algorithms.CryptoTestHelper.assertSignaturePresent;
+import static com.auth0.jwt.algorithms.CryptoTestHelper.assertSignatureValue;
 import static org.hamcrest.Matchers.*;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.Assert.assertArrayEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
@@ -28,91 +33,106 @@ public class HMACAlgorithmTest {
     // Verify
 
     @Test
-    public void shouldGetStringBytes() throws Exception {
+    public void shouldGetStringBytes() {
         String text = "abcdef123456!@#$%^";
-        byte[] expectedBytes = text.getBytes("UTF-8");
-        assertTrue(Arrays.equals(expectedBytes, HMACAlgorithm.getSecretBytes(text)));
+        byte[] expectedBytes = text.getBytes(StandardCharsets.UTF_8);
+        assertArrayEquals(expectedBytes, HMACAlgorithm.getSecretBytes(text));
     }
 
     @Test
-    public void shouldPassHMAC256Verification() throws Exception {
+    public void shouldCopyTheReceivedSecretArray() {
+        String jwt = "eyJhbGciOiJIUzI1NiIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJhdXRoMCJ9.mZ0m_N1J4PgeqWmi903JuUoDRZDBPB7HwkS4nVyWH1M";
+        byte[] secretArray = "secret".getBytes(Charset.defaultCharset());
+        Algorithm algorithmString = Algorithm.HMAC256(secretArray);
+
+        DecodedJWT decoded = JWT.decode(jwt);
+        algorithmString.verify(decoded);
+        secretArray[0] = secretArray[1];
+        algorithmString.verify(decoded);
+    }
+
+    @Test
+    public void shouldPassHMAC256Verification() {
         String jwt = "eyJhbGciOiJIUzI1NiIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJhdXRoMCJ9.mZ0m_N1J4PgeqWmi903JuUoDRZDBPB7HwkS4nVyWH1M";
         Algorithm algorithmString = Algorithm.HMAC256("secret");
         Algorithm algorithmBytes = Algorithm.HMAC256("secret".getBytes(StandardCharsets.UTF_8));
-        AlgorithmUtils.verify(algorithmString, jwt);
-        AlgorithmUtils.verify(algorithmBytes, jwt);
+        DecodedJWT decoded = JWT.decode(jwt);
+        algorithmString.verify(decoded);
+        algorithmBytes.verify(decoded);
     }
 
     @Test
-    public void shouldFailHMAC256VerificationWithInvalidSecretString() throws Exception {
+    public void shouldFailHMAC256VerificationWithInvalidSecretString() {
         exception.expect(SignatureVerificationException.class);
         exception.expectMessage("The Token's Signature resulted invalid when verified using the Algorithm: HmacSHA256");
         String jwt = "eyJhbGciOiJIUzI1NiIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJhdXRoMCJ9.mZ0m_N1J4PgeqWmi903JuUoDRZDBPB7HwkS4nVyWH1M";
         Algorithm algorithm = Algorithm.HMAC256("not_real_secret");
-        AlgorithmUtils.verify(algorithm, jwt);
+        algorithm.verify(JWT.decode(jwt));
     }
 
     @Test
-    public void shouldFailHMAC256VerificationWithInvalidSecretBytes() throws Exception {
+    public void shouldFailHMAC256VerificationWithInvalidSecretBytes() {
         exception.expect(SignatureVerificationException.class);
         exception.expectMessage("The Token's Signature resulted invalid when verified using the Algorithm: HmacSHA256");
         String jwt = "eyJhbGciOiJIUzI1NiIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJhdXRoMCJ9.mZ0m_N1J4PgeqWmi903JuUoDRZDBPB7HwkS4nVyWH1M";
         Algorithm algorithm = Algorithm.HMAC256("not_real_secret".getBytes(StandardCharsets.UTF_8));
-        AlgorithmUtils.verify(algorithm, jwt);
+        algorithm.verify(JWT.decode(jwt));
     }
 
     @Test
-    public void shouldPassHMAC384Verification() throws Exception {
+    public void shouldPassHMAC384Verification() {
         String jwt = "eyJhbGciOiJIUzM4NCIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJhdXRoMCJ9.uztpK_wUMYJhrRv8SV-1LU4aPnwl-EM1q-wJnqgyb5DHoDteP6lN_gE1xnZJH5vw";
         Algorithm algorithmString = Algorithm.HMAC384("secret");
         Algorithm algorithmBytes = Algorithm.HMAC384("secret".getBytes(StandardCharsets.UTF_8));
-        AlgorithmUtils.verify(algorithmString, jwt);
-        AlgorithmUtils.verify(algorithmBytes, jwt);
+        DecodedJWT decoded = JWT.decode(jwt);
+        algorithmString.verify(decoded);
+        algorithmBytes.verify(decoded);
     }
 
     @Test
-    public void shouldFailHMAC384VerificationWithInvalidSecretString() throws Exception {
+    public void shouldFailHMAC384VerificationWithInvalidSecretString() {
         exception.expect(SignatureVerificationException.class);
         exception.expectMessage("The Token's Signature resulted invalid when verified using the Algorithm: HmacSHA384");
         String jwt = "eyJhbGciOiJIUzM4NCIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJhdXRoMCJ9.uztpK_wUMYJhrRv8SV-1LU4aPnwl-EM1q-wJnqgyb5DHoDteP6lN_gE1xnZJH5vw";
         Algorithm algorithm = Algorithm.HMAC384("not_real_secret");
-        AlgorithmUtils.verify(algorithm, jwt);
+        algorithm.verify(JWT.decode(jwt));
     }
 
     @Test
-    public void shouldFailHMAC384VerificationWithInvalidSecretBytes() throws Exception {
+    public void shouldFailHMAC384VerificationWithInvalidSecretBytes() {
         exception.expect(SignatureVerificationException.class);
         exception.expectMessage("The Token's Signature resulted invalid when verified using the Algorithm: HmacSHA384");
         String jwt = "eyJhbGciOiJIUzM4NCIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJhdXRoMCJ9.uztpK_wUMYJhrRv8SV-1LU4aPnwl-EM1q-wJnqgyb5DHoDteP6lN_gE1xnZJH5vw";
         Algorithm algorithm = Algorithm.HMAC384("not_real_secret".getBytes(StandardCharsets.UTF_8));
-        AlgorithmUtils.verify(algorithm, jwt);
+        algorithm.verify(JWT.decode(jwt));
     }
 
     @Test
-    public void shouldPassHMAC512Verification() throws Exception {
+    public void shouldPassHMAC512Verification() {
         String jwt = "eyJhbGciOiJIUzUxMiIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJhdXRoMCJ9.VUo2Z9SWDV-XcOc_Hr6Lff3vl7L9e5Vb8ThXpmGDFjHxe3Dr1ZBmUChYF-xVA7cAdX1P_D4ZCUcsv3IefpVaJw";
         Algorithm algorithmString = Algorithm.HMAC512("secret");
         Algorithm algorithmBytes = Algorithm.HMAC512("secret".getBytes(StandardCharsets.UTF_8));
-        AlgorithmUtils.verify(algorithmString, jwt);
-        AlgorithmUtils.verify(algorithmBytes, jwt);
+        DecodedJWT decoded = JWT.decode(jwt);
+        algorithmString.verify(decoded);
+        algorithmBytes.verify(decoded);
     }
 
     @Test
-    public void shouldFailHMAC512VerificationWithInvalidSecretString() throws Exception {
+    public void shouldFailHMAC512VerificationWithInvalidSecretString() {
         exception.expect(SignatureVerificationException.class);
         exception.expectMessage("The Token's Signature resulted invalid when verified using the Algorithm: HmacSHA512");
         String jwt = "eyJhbGciOiJIUzUxMiIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJhdXRoMCJ9.VUo2Z9SWDV-XcOc_Hr6Lff3vl7L9e5Vb8ThXpmGDFjHxe3Dr1ZBmUChYF-xVA7cAdX1P_D4ZCUcsv3IefpVaJw";
         Algorithm algorithm = Algorithm.HMAC512("not_real_secret");
-        AlgorithmUtils.verify(algorithm, jwt);
+        algorithm.verify(JWT.decode(jwt));
     }
 
     @Test
-    public void shouldFailHMAC512VerificationWithInvalidSecretBytes() throws Exception {
+    public void shouldFailHMAC512VerificationWithInvalidSecretBytes() {
         exception.expect(SignatureVerificationException.class);
         exception.expectMessage("The Token's Signature resulted invalid when verified using the Algorithm: HmacSHA512");
         String jwt = "eyJhbGciOiJIUzUxMiIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJhdXRoMCJ9.VUo2Z9SWDV-XcOc_Hr6Lff3vl7L9e5Vb8ThXpmGDFjHxe3Dr1ZBmUChYF-xVA7cAdX1P_D4ZCUcsv3IefpVaJw";
         Algorithm algorithm = Algorithm.HMAC512("not_real_secret".getBytes(StandardCharsets.UTF_8));
-        AlgorithmUtils.verify(algorithm, jwt);
+        algorithm.verify(JWT.decode(jwt));
     }
 
     @Test
@@ -122,12 +142,12 @@ public class HMACAlgorithmTest {
         exception.expectCause(isA(NoSuchAlgorithmException.class));
 
         CryptoHelper crypto = mock(CryptoHelper.class);
-        when(crypto.verifySignatureFor(anyString(), any(byte[].class), any(byte[].class), any(byte[].class)))
+        when(crypto.verifySignatureFor(anyString(), any(byte[].class), any(String.class), any(String.class), any(byte[].class)))
                 .thenThrow(NoSuchAlgorithmException.class);
 
         Algorithm algorithm = new HMACAlgorithm(crypto, "some-alg", "some-algorithm", "secret".getBytes(StandardCharsets.UTF_8));
         String jwt = "eyJhbGciOiJIUzI1NiIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJhdXRoMCJ9.mZ0m_N1J4PgeqWmi903JuUoDRZDBPB7HwkS4nVyWH1M";
-        AlgorithmUtils.verify(algorithm, jwt);
+        algorithm.verify(JWT.decode(jwt));
     }
 
     @Test
@@ -137,12 +157,12 @@ public class HMACAlgorithmTest {
         exception.expectCause(isA(InvalidKeyException.class));
 
         CryptoHelper crypto = mock(CryptoHelper.class);
-        when(crypto.verifySignatureFor(anyString(), any(byte[].class), any(byte[].class), any(byte[].class)))
+        when(crypto.verifySignatureFor(anyString(), any(byte[].class), any(String.class), any(String.class), any(byte[].class)))
                 .thenThrow(InvalidKeyException.class);
 
         Algorithm algorithm = new HMACAlgorithm(crypto, "some-alg", "some-algorithm", "secret".getBytes(StandardCharsets.UTF_8));
         String jwt = "eyJhbGciOiJIUzI1NiIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJhdXRoMCJ9.mZ0m_N1J4PgeqWmi903JuUoDRZDBPB7HwkS4nVyWH1M";
-        AlgorithmUtils.verify(algorithm, jwt);
+        algorithm.verify(JWT.decode(jwt));
     }
 
     // Sign
@@ -153,81 +173,75 @@ public class HMACAlgorithmTest {
     private static final String auth0IssPayload = "eyJpc3MiOiJhdXRoMCJ9";
 
     @Test
-    public void shouldDoHMAC256SigningWithBytes() throws Exception {
+    public void shouldDoHMAC256SigningWithBytes() {
         Algorithm algorithm = Algorithm.HMAC256("secret".getBytes(StandardCharsets.UTF_8));
-        byte[] contentBytes = String.format("%s.%s", HS256Header, auth0IssPayload).getBytes(StandardCharsets.UTF_8);
-        byte[] signatureBytes = algorithm.sign(contentBytes);
-        String signature = Base64.encodeBase64URLSafeString(signatureBytes);
+
+        String jwt = asJWT(algorithm, HS256Header, auth0IssPayload);
         String expectedSignature = "s69x7Mmu4JqwmdxiK6sesALO7tcedbFsKEEITUxw9ho";
 
-        assertThat(signatureBytes, is(notNullValue()));
-        assertThat(signature, is(expectedSignature));
-        algorithm.verify(contentBytes, signatureBytes);
+        assertSignaturePresent(jwt);
+        assertSignatureValue(jwt, expectedSignature);
+        algorithm.verify(JWT.decode(jwt));
     }
 
     @Test
-    public void shouldDoHMAC384SigningWithBytes() throws Exception {
+    public void shouldDoHMAC384SigningWithBytes() {
         Algorithm algorithm = Algorithm.HMAC384("secret".getBytes(StandardCharsets.UTF_8));
-        byte[] contentBytes = String.format("%s.%s", HS384Header, auth0IssPayload).getBytes(StandardCharsets.UTF_8);
-        byte[] signatureBytes = algorithm.sign(contentBytes);
-        String signature = Base64.encodeBase64URLSafeString(signatureBytes);
+
+        String jwt = asJWT(algorithm, HS384Header, auth0IssPayload);
         String expectedSignature = "4-y2Gxz_foN0jAOFimmBPF7DWxf4AsjM20zxNkHg8Zah5Q64G42P9GfjmUp4Hldt";
 
-        assertThat(signatureBytes, is(notNullValue()));
-        assertThat(signature, is(expectedSignature));
-        algorithm.verify(contentBytes, signatureBytes);
+        assertSignaturePresent(jwt);
+        assertSignatureValue(jwt, expectedSignature);
+        algorithm.verify(JWT.decode(jwt));
     }
 
     @Test
-    public void shouldDoHMAC512SigningWithBytes() throws Exception {
+    public void shouldDoHMAC512SigningWithBytes() {
         Algorithm algorithm = Algorithm.HMAC512("secret".getBytes(StandardCharsets.UTF_8));
-        byte[] contentBytes = String.format("%s.%s", HS512Header, auth0IssPayload).getBytes(StandardCharsets.UTF_8);
-        byte[] signatureBytes = algorithm.sign(contentBytes);
-        String signature = Base64.encodeBase64URLSafeString(signatureBytes);
+
+        String jwt = asJWT(algorithm, HS512Header, auth0IssPayload);
         String expectedSignature = "OXWyxmf-VcVo8viOiTFfLaEy6mrQqLEos5R82Xsx8mtFxQadJAQ1aVniIWN8qT2GNE_pMQPcdzk4x7Cqxsp1dw";
 
-        assertThat(signatureBytes, is(notNullValue()));
-        assertThat(signature, is(expectedSignature));
-        algorithm.verify(contentBytes, signatureBytes);
+        assertSignaturePresent(jwt);
+        assertSignatureValue(jwt, expectedSignature);
+        algorithm.verify(JWT.decode(jwt));
     }
 
     @Test
-    public void shouldDoHMAC256SigningWithString() throws Exception {
+    public void shouldDoHMAC256SigningWithString() {
         Algorithm algorithm = Algorithm.HMAC256("secret");
-        byte[] contentBytes = String.format("%s.%s", HS256Header, auth0IssPayload).getBytes(StandardCharsets.UTF_8);
-        byte[] signatureBytes = algorithm.sign(contentBytes);
-        String signature = Base64.encodeBase64URLSafeString(signatureBytes);
+
+        String jwt = asJWT(algorithm, HS256Header, auth0IssPayload);
         String expectedSignature = "s69x7Mmu4JqwmdxiK6sesALO7tcedbFsKEEITUxw9ho";
 
-        assertThat(signatureBytes, is(notNullValue()));
-        assertThat(signature, is(expectedSignature));
-        algorithm.verify(contentBytes, signatureBytes);
+        assertSignaturePresent(jwt);
+        assertSignatureValue(jwt, expectedSignature);
+        algorithm.verify(JWT.decode(jwt));
     }
 
     @Test
-    public void shouldDoHMAC384SigningWithString() throws Exception {
+    public void shouldDoHMAC384SigningWithString() {
         Algorithm algorithm = Algorithm.HMAC384("secret");
-        byte[] contentBytes = String.format("%s.%s", HS384Header, auth0IssPayload).getBytes(StandardCharsets.UTF_8);
-        byte[] signatureBytes = algorithm.sign(contentBytes);
-        String signature = Base64.encodeBase64URLSafeString(signatureBytes);
+
+        String jwt = asJWT(algorithm, HS384Header, auth0IssPayload);
         String expectedSignature = "4-y2Gxz_foN0jAOFimmBPF7DWxf4AsjM20zxNkHg8Zah5Q64G42P9GfjmUp4Hldt";
 
-        assertThat(signatureBytes, is(notNullValue()));
-        assertThat(signature, is(expectedSignature));
-        algorithm.verify(contentBytes, signatureBytes);
+        assertSignaturePresent(jwt);
+        assertSignatureValue(jwt, expectedSignature);
+        algorithm.verify(JWT.decode(jwt));
     }
 
     @Test
-    public void shouldDoHMAC512SigningWithString() throws Exception {
+    public void shouldDoHMAC512SigningWithString() {
         Algorithm algorithm = Algorithm.HMAC512("secret");
-        byte[] contentBytes = String.format("%s.%s", HS512Header, auth0IssPayload).getBytes(StandardCharsets.UTF_8);
-        byte[] signatureBytes = algorithm.sign(contentBytes);
-        String signature = Base64.encodeBase64URLSafeString(signatureBytes);
+
+        String jwt = asJWT(algorithm ,HS512Header, auth0IssPayload);
         String expectedSignature = "OXWyxmf-VcVo8viOiTFfLaEy6mrQqLEos5R82Xsx8mtFxQadJAQ1aVniIWN8qT2GNE_pMQPcdzk4x7Cqxsp1dw";
 
-        assertThat(signatureBytes, is(notNullValue()));
-        assertThat(signature, is(expectedSignature));
-        algorithm.verify(contentBytes, signatureBytes);
+        assertSignaturePresent(jwt);
+        assertSignatureValue(jwt, expectedSignature);
+        algorithm.verify(JWT.decode(jwt));
     }
 
     @Test
@@ -237,11 +251,11 @@ public class HMACAlgorithmTest {
         exception.expectCause(isA(NoSuchAlgorithmException.class));
 
         CryptoHelper crypto = mock(CryptoHelper.class);
-        when(crypto.createSignatureFor(anyString(), any(byte[].class), any(byte[].class)))
+        when(crypto.createSignatureFor(anyString(), any(byte[].class), any(byte[].class), any(byte[].class)))
                 .thenThrow(NoSuchAlgorithmException.class);
 
         Algorithm algorithm = new HMACAlgorithm(crypto, "some-alg", "some-algorithm", "secret".getBytes(StandardCharsets.UTF_8));
-        algorithm.sign(new byte[0]);
+        algorithm.sign(new byte[0], new byte[0]);
     }
 
     @Test
@@ -251,11 +265,45 @@ public class HMACAlgorithmTest {
         exception.expectCause(isA(InvalidKeyException.class));
 
         CryptoHelper crypto = mock(CryptoHelper.class);
-        when(crypto.createSignatureFor(anyString(), any(byte[].class), any(byte[].class)))
+        when(crypto.createSignatureFor(anyString(), any(byte[].class), any(byte[].class), any(byte[].class)))
                 .thenThrow(InvalidKeyException.class);
 
         Algorithm algorithm = new HMACAlgorithm(crypto, "some-alg", "some-algorithm", "secret".getBytes(StandardCharsets.UTF_8));
-        algorithm.sign(new byte[0]);
+        algorithm.sign(new byte[0], new byte[0]);
     }
 
+    @Test
+    public void shouldReturnNullSigningKeyId() {
+        assertThat(Algorithm.HMAC256("secret").getSigningKeyId(), is(nullValue()));
+    }
+
+    @Test
+    public void shouldBeEqualSignatureMethodResults() throws Exception {
+        Algorithm algorithm = Algorithm.HMAC256("secret");
+
+        byte[] header = new byte[]{0x00, 0x01, 0x02};
+        byte[] payload = new byte[]{0x04, 0x05, 0x06};
+
+        ByteArrayOutputStream bout = new ByteArrayOutputStream();
+        bout.write(header);
+        bout.write('.');
+        bout.write(payload);
+
+        assertThat(algorithm.sign(bout.toByteArray()), is(algorithm.sign(header, payload)));
+    }
+
+
+    @Test
+    public void shouldThrowWhenSignatureNotValidBase64() throws Exception {
+        exception.expect(SignatureVerificationException.class);
+        exception.expectCause(isA(IllegalArgumentException.class));
+
+        CryptoHelper crypto = mock(CryptoHelper.class);
+        when(crypto.verifySignatureFor(anyString(), any(byte[].class), any(String.class), any(String.class), any(byte[].class)))
+                .thenThrow(NoSuchAlgorithmException.class);
+
+        Algorithm algorithm = new HMACAlgorithm(crypto, "some-alg", "some-algorithm", "secret".getBytes(StandardCharsets.UTF_8));
+        String jwt = "eyJhbGciOiJIUzI1NiIsImN0eSI6IkpXVCJ9.eyJpc3MiOiJhdXRoMCJ9.mZ0m_N1J4PgeqWm+i903JuUoDRZDBPB7HwkS4nVyWH1M";
+        algorithm.verify(JWT.decode(jwt));
+    }
 }
