@@ -1,6 +1,8 @@
 package com.auth0.jwt.impl;
 
 import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.ObjectReader;
 import com.fasterxml.jackson.databind.node.NullNode;
 import com.fasterxml.jackson.databind.node.TextNode;
 import org.hamcrest.collection.IsMapContaining;
@@ -18,36 +20,38 @@ public class BasicHeaderTest {
 
     @Rule
     public ExpectedException exception = ExpectedException.none();
+    
+    private ObjectReader objectReader = new ObjectMapper().reader();
 
     @SuppressWarnings("Convert2Diamond")
     @Test
-    public void shouldHaveUnmodifiableTreeWhenInstantiatedWithNonNullTree() throws Exception {
+    public void shouldHaveUnmodifiableTreeWhenInstantiatedWithNonNullTree() {
         exception.expect(UnsupportedOperationException.class);
-        BasicHeader header = new BasicHeader(null, null, null, null, new HashMap<String, JsonNode>());
+        BasicHeader header = new BasicHeader(null, null, null, null, new HashMap<String, JsonNode>(), objectReader);
         header.getTree().put("something", null);
     }
 
     @Test
-    public void shouldHaveUnmodifiableTreeWhenInstantiatedWithNullTree() throws Exception {
+    public void shouldHaveUnmodifiableTreeWhenInstantiatedWithNullTree() {
         exception.expect(UnsupportedOperationException.class);
-        BasicHeader header = new BasicHeader(null, null, null, null, null);
+        BasicHeader header = new BasicHeader(null, null, null, null, null, objectReader);
         header.getTree().put("something", null);
     }
 
     @Test
-    public void shouldHaveTree() throws Exception {
+    public void shouldHaveTree() {
         HashMap<String, JsonNode> map = new HashMap<>();
         JsonNode node = NullNode.getInstance();
         map.put("key", node);
-        BasicHeader header = new BasicHeader(null, null, null, null, map);
+        BasicHeader header = new BasicHeader(null, null, null, null, map, objectReader);
 
         assertThat(header.getTree(), is(notNullValue()));
         assertThat(header.getTree(), is(IsMapContaining.hasEntry("key", node)));
     }
 
     @Test
-    public void shouldGetAlgorithm() throws Exception {
-        BasicHeader header = new BasicHeader("HS256", null, null, null, null);
+    public void shouldGetAlgorithm() {
+        BasicHeader header = new BasicHeader("HS256", null, null, null, null, objectReader);
 
         assertThat(header, is(notNullValue()));
         assertThat(header.getAlgorithm(), is(notNullValue()));
@@ -55,16 +59,16 @@ public class BasicHeaderTest {
     }
 
     @Test
-    public void shouldGetNullAlgorithmIfMissing() throws Exception {
-        BasicHeader header = new BasicHeader(null, null, null, null, null);
+    public void shouldGetNullAlgorithmIfMissing() {
+        BasicHeader header = new BasicHeader(null, null, null, null, null, objectReader);
 
         assertThat(header, is(notNullValue()));
         assertThat(header.getAlgorithm(), is(nullValue()));
     }
 
     @Test
-    public void shouldGetType() throws Exception {
-        BasicHeader header = new BasicHeader(null, "jwt", null, null, null);
+    public void shouldGetType() {
+        BasicHeader header = new BasicHeader(null, "jwt", null, null, null, objectReader);
 
         assertThat(header, is(notNullValue()));
         assertThat(header.getType(), is(notNullValue()));
@@ -72,16 +76,16 @@ public class BasicHeaderTest {
     }
 
     @Test
-    public void shouldGetNullTypeIfMissing() throws Exception {
-        BasicHeader header = new BasicHeader(null, null, null, null, null);
+    public void shouldGetNullTypeIfMissing() {
+        BasicHeader header = new BasicHeader(null, null, null, null, null, objectReader);
 
         assertThat(header, is(notNullValue()));
         assertThat(header.getType(), is(nullValue()));
     }
 
     @Test
-    public void shouldGetContentType() throws Exception {
-        BasicHeader header = new BasicHeader(null, null, "content", null, null);
+    public void shouldGetContentType() {
+        BasicHeader header = new BasicHeader(null, null, "content", null, null, objectReader);
 
         assertThat(header, is(notNullValue()));
         assertThat(header.getContentType(), is(notNullValue()));
@@ -89,16 +93,16 @@ public class BasicHeaderTest {
     }
 
     @Test
-    public void shouldGetNullContentTypeIfMissing() throws Exception {
-        BasicHeader header = new BasicHeader(null, null, null, null, null);
+    public void shouldGetNullContentTypeIfMissing() {
+        BasicHeader header = new BasicHeader(null, null, null, null, null, objectReader);
 
         assertThat(header, is(notNullValue()));
         assertThat(header.getContentType(), is(nullValue()));
     }
 
     @Test
-    public void shouldGetKeyId() throws Exception {
-        BasicHeader header = new BasicHeader(null, null, null, "key", null);
+    public void shouldGetKeyId() {
+        BasicHeader header = new BasicHeader(null, null, null, "key", null, objectReader);
 
         assertThat(header, is(notNullValue()));
         assertThat(header.getKeyId(), is(notNullValue()));
@@ -106,18 +110,18 @@ public class BasicHeaderTest {
     }
 
     @Test
-    public void shouldGetNullKeyIdIfMissing() throws Exception {
-        BasicHeader header = new BasicHeader(null, null, null, null, null);
+    public void shouldGetNullKeyIdIfMissing() {
+        BasicHeader header = new BasicHeader(null, null, null, null, null, objectReader);
 
         assertThat(header, is(notNullValue()));
         assertThat(header.getKeyId(), is(nullValue()));
     }
 
     @Test
-    public void shouldGetExtraClaim() throws Exception {
+    public void shouldGetExtraClaim() {
         Map<String, JsonNode> tree = new HashMap<>();
         tree.put("extraClaim", new TextNode("extraValue"));
-        BasicHeader header = new BasicHeader(null, null, null, null, tree);
+        BasicHeader header = new BasicHeader(null, null, null, null, tree, objectReader);
 
         assertThat(header, is(notNullValue()));
         assertThat(header.getHeaderClaim("extraClaim"), is(instanceOf(JsonNodeClaim.class)));
@@ -125,12 +129,13 @@ public class BasicHeaderTest {
     }
 
     @Test
-    public void shouldGetNotNullExtraClaimIfMissing() throws Exception {
+    public void shouldGetNotNullExtraClaimIfMissing() {
         Map<String, JsonNode> tree = new HashMap<>();
-        BasicHeader header = new BasicHeader(null, null, null, null, tree);
+        BasicHeader header = new BasicHeader(null, null, null, null, tree, objectReader);
 
         assertThat(header, is(notNullValue()));
         assertThat(header.getHeaderClaim("missing"), is(notNullValue()));
-        assertThat(header.getHeaderClaim("missing"), is(instanceOf(NullClaim.class)));
+        assertThat(header.getHeaderClaim("missing").isMissing(), is(true));
+        assertThat(header.getHeaderClaim("missing").isNull(), is(false));
     }
 }
