@@ -7,25 +7,24 @@ import java.util.HashMap;
 import java.util.Map;
 
 import static org.hamcrest.Matchers.*;
-import static org.junit.Assert.assertThat;
+import static org.hamcrest.MatcherAssert.assertThat;
 
-public class ClaimsHolderTest {
+public class PayloadClaimsHolderTest {
 
-    @SuppressWarnings("RedundantCast")
     @Test
-    public void shouldGetClaims() throws Exception {
+    public void shouldGetClaims() {
         HashMap<String, Object> claims = new HashMap<>();
         claims.put("iss", "auth0");
-        ClaimsHolder holder = new ClaimsHolder(claims);
+        ClaimsHolder holder = new PayloadClaimsHolder(claims);
         assertThat(holder, is(notNullValue()));
         assertThat(holder.getClaims(), is(notNullValue()));
         assertThat(holder.getClaims(), is(instanceOf(Map.class)));
-        assertThat(holder.getClaims(), is(IsMapContaining.hasEntry("iss", (Object) "auth0")));
+        assertThat(holder.getClaims(), is(IsMapContaining.hasEntry("iss", "auth0")));
     }
 
     @Test
-    public void shouldGetNotNullClaims() throws Exception {
-        ClaimsHolder holder = new ClaimsHolder(null);
+    public void shouldGetNotNullClaims() {
+        ClaimsHolder holder = new PayloadClaimsHolder(null);
         assertThat(holder, is(notNullValue()));
         assertThat(holder.getClaims(), is(notNullValue()));
         assertThat(holder.getClaims(), is(instanceOf(Map.class)));
